# Invoke the locale plugin rake tasks directly