We have lots of plugins and there should be no need to shell out and constantly
load the rails environment and application over and over again.

I've verified that we're generating the same po and pot files with direct rake
task invocation but we save several minutes.

Before this commit:

```
bundle exec rake locale:update_all  248.95s user 83.08s system 96% cpu 5:43.01 total
```

After this commit:

```
bundle exec rake locale:update_all  78.93s user 8.90s system 97% cpu 1:29.70 total
```

Here is a book describing how I tested this:

Po/pot files are not updated each time we change strings throughout our codebase so they become stale quickly.  Therefore we need to generate a baseline on master before we can evaluate if any changes we make to the po / pot generation, for refacoring or performance improvements, accidentally changed the outputted files.

Here's how I verify this.

### Capture the baseline message catalogs on master

1) git checkout master
2) git checkout -b baseline
3) bundle exec rake locale:update_all # generates the updated message catalogs for master
4) git add locale
5) git commit -m "Add baseline master catalogs"
6) Do step 3 again and check git diff # this should show only a few timestamp fields updated

### Start a "changes" branch

1) git checkout master # use the same base branch as the baseline to avoid plugin/other changes
2) git checkout -b "changes" # Now, make your code changes
3) bundle exec rake locale:update_all # verify locale/ has changes to the po/pot
4) git add locale
5) git commit -m "my changes"
6) Compare the locale/ directory from this new branch to the baseline:
6b) git diff baseline HEAD locale/ # this should show only a few timestamp fields updated
7) Do step 3 again # this should show only a few timestamp fields updated

If you are just refactoring code or making performance improvements, 6b should show no differences from the baseline.