@jrafanie 2022-06-24
Ok, I've simplified this PR down to the basics and added a 📖 on how I tested this to ensure I didn't break how we generate message catalogs.

Basically, this branch and master will generate the same message catalogs with `rake locale:update_all` but this branch will take over 4 minutes left.